requirements.txt
# Trust-Drift Pipeline - Requirements
# ====================================
# Install with: pip install -r requirements.txt

# Core ML/Data
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# API/Web
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0

# Configuration
pyyaml>=6.0

# Development/Optional
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
